+++ zetaclient/chains/base/observer.go
@@ -60,6 +60,10 @@ type Observer struct {
 	// lastTxScanned is the last transaction hash scanned by the observer
 	lastTxScanned string
 
+	// auxStringMap is a key-value map to store any auxiliary string values used by the observer
+	// it is now only used by Sui observer to store old/new Sui gateway inbound cursors
+	auxStringMap map[string]string
+
 	blockCache *lru.Cache
 
 	// db is the database to persist data
@@ -106,6 +110,7 @@ func NewObserver(
 		lastBlock:             0,
 		lastBlockScanned:      0,
 		lastTxScanned:         "",
+		auxStringMap:          make(map[string]string),
 		ts:                    ts,
 		db:                    database,
 		blockCache:            blockCache,
@@ -249,6 +254,8 @@ func (ob *Observer) WithLastBlockScanned(blockNumber uint64, forceResetLastScann
 
 // LastTxScanned get last transaction scanned.
 func (ob *Observer) LastTxScanned() string {
+	ob.mu.Lock()
+	defer ob.mu.Unlock()
 	return ob.lastTxScanned
 }
 
@@ -257,7 +264,11 @@ func (ob *Observer) WithLastTxScanned(txHash string) *Observer {
 	ob.mu.Lock()
 	defer ob.mu.Unlock()
 
+	if ob.lastTxScanned == "" {
+		ob.logger.Chain.Info().Str("tx", txHash).Msg("initializing last tx scanned")
+	}
 	ob.lastTxScanned = txHash
+
 	return ob
 }
 
@@ -301,7 +312,7 @@ func (ob *Observer) StopChannel() chan struct{} {
 // LoadLastBlockScanned loads last scanned block from environment variable or from database.
 // The last scanned block is the height from which the observer should continue scanning.
 func (ob *Observer) LoadLastBlockScanned() error {
-	logger := ob.logger.Chain.With().Str(logs.FieldMethod, "LoadLastBlockScanned").Logger()
+	logger := ob.logger.Chain
 
 	// get environment variable
 	envvar := EnvVarLatestBlockByChain(ob.chain)
@@ -373,7 +384,7 @@ func (ob *Observer) ReadLastBlockScannedFromDB() (uint64, error) {
 // LoadLastTxScanned loads last scanned tx from environment variable or from database.
 // The last scanned tx is the tx hash from which the observer should continue scanning.
 func (ob *Observer) LoadLastTxScanned() {
-	logger := ob.logger.Chain.With().Str(logs.FieldMethod, "LoadLastTxScanned").Logger()
+	logger := ob.logger.Chain
 
 	// get environment variable
 	envvar := EnvVarLatestTxByChain(ob.chain)
@@ -425,6 +436,71 @@ func (ob *Observer) ReadLastTxScannedFromDB() (string, error) {
 	return lastTx.Hash, nil
 }
 
+// GetAuxString get any auxiliary string data by key
+func (ob *Observer) GetAuxString(key string) string {
+	ob.mu.Lock()
+	defer ob.mu.Unlock()
+	return ob.auxStringMap[key]
+}
+
+// WithAuxString set any auxiliary string data by key
+func (ob *Observer) WithAuxString(key, value string) *Observer {
+	ob.mu.Lock()
+	defer ob.mu.Unlock()
+
+	if ob.auxStringMap[key] == "" {
+		ob.logger.Chain.Info().Str("key", key).Str("value", value).Msg("initializing auxiliary string value")
+	}
+	ob.auxStringMap[key] = value
+
+	return ob
+}
+
+// WriteAuxStringToDB writes the auxiliary string data to the database.
+func (ob *Observer) WriteAuxStringToDB(key, value string) error {
+	// create new record if not found
+	var existingRecord clienttypes.AuxStringSQLType
+	if err := ob.db.Client().Where("key_name = ?", key).First(&existingRecord).Error; err != nil {
+		return ob.db.Client().Create(clienttypes.ToAuxStringSQLType(key, value)).Error
+	}
+
+	// record exists, update it
+	return ob.db.Client().Model(&existingRecord).Update("value", value).Error
+}
+
+// LoadAuxString loads auxiliary string data from environment variable or from database.
+func (ob *Observer) LoadAuxString(key string) {
+	// get environment variable
+	envvar := EnvVarLatestAuxStringByChain(ob.chain, key)
+	value := os.Getenv(envvar)
+
+	// load from environment variable if set
+	if value != "" {
+		ob.logger.Chain.Info().Str("envvar", envvar).Str("value", value).Msg("environment variable is set")
+		ob.WithAuxString(key, value)
+		return
+	}
+
+	// load from DB otherwise
+	value, err := ob.ReadAuxStringFromDB(key)
+	if err != nil {
+		// if not found, let the concrete chain observer decide where to start
+		chainID := ob.chain.ChainId
+		ob.logger.Chain.Info().Int64(logs.FieldChain, chainID).Str("key", key).Msg("string value not found in db")
+		return
+	}
+	ob.WithAuxString(key, value)
+}
+
+// ReadAuxStringFromDB reads the auxiliary string data from the database.
+func (ob *Observer) ReadAuxStringFromDB(key string) (string, error) {
+	var record clienttypes.AuxStringSQLType
+	if err := ob.db.Client().Where("key_name = ?", key).First(&record).Error; err != nil {
+		return "", err
+	}
+	return record.Value, nil
+}
+
 // PostVoteInbound posts a vote for the given vote message and returns the ballot.
 func (ob *Observer) PostVoteInbound(
 	ctx context.Context,
@@ -439,10 +515,9 @@ func (ob *Observer) PostVoteInbound(
 	)
 
 	logger := ob.logger.Inbound.With().
-		Str(logs.FieldMethod, "PostVoteInbound").
 		Str(logs.FieldTx, txHash).
 		Stringer(logs.FieldCoinType, coinType).
-		Stringer(logs.FieldConfirmationMode, msg.ConfirmationMode).
+		Stringer("confirmation_mode", msg.ConfirmationMode).
 		Logger()
 
 	cctxIndex := msg.Digest()
@@ -476,7 +551,7 @@ func (ob *Observer) PostVoteInbound(
 
 	logger = logger.With().
 		Str(logs.FieldZetaTx, zetaHash).
-		Str(logs.FieldBallot, ballot).
+		Str(logs.FieldBallotIndex, ballot).
 		Logger()
 
 	switch {
@@ -513,9 +588,8 @@ func (ob *Observer) handleMonitoringError(
 		if monitorErr.Err != nil {
 			logger.Error().
 				Err(monitorErr).
-				Str(logs.FieldMethod, "handleMonitoringError").
 				Str(logs.FieldZetaTx, monitorErr.ZetaTxHash).
-				Str(logs.FieldBallot, monitorErr.BallotIndex).
+				Str(logs.FieldBallotIndex, monitorErr.BallotIndex).
 				Uint64(logs.FieldBlock, monitorErr.InboundBlockHeight).
 				Msg("error monitoring vote transaction")
 
@@ -543,11 +617,16 @@ func EnvVarLatestTxByChain(chain chains.Chain) string {
 	return fmt.Sprintf("CHAIN_%d_SCAN_FROM_TX", chain.ChainId)
 }
 
+// EnvVarLatestAuxStringByChain returns the environment variable for auxiliary string data by chain for the given key.
+func EnvVarLatestAuxStringByChain(chain chains.Chain, key string) string {
+	return fmt.Sprintf("CHAIN_%d_AUX_STRING_%s", chain.ChainId, key)
+}
+
 func newObserverLogger(chain chains.Chain, logger Logger) ObserverLogger {
 	withLogFields := func(l zerolog.Logger) zerolog.Logger {
 		return l.With().
 			Int64(logs.FieldChain, chain.ChainId).
-			Str(logs.FieldChainNetwork, chain.Network.String()).
+			Stringer(logs.FieldNetwork, chain.Network).
 			Logger()
 	}
 
@@ -558,8 +637,6 @@ func newObserverLogger(chain chains.Chain, logger Logger) ObserverLogger {
 		Chain:      log,
 		Inbound:    log.With().Str(logs.FieldModule, logs.ModNameInbound).Logger(),
 		Outbound:   log.With().Str(logs.FieldModule, logs.ModNameOutbound).Logger(),
-		GasPrice:   log.With().Str(logs.FieldModule, logs.ModNameGasPrice).Logger(),
-		Headers:    log.With().Str(logs.FieldModule, logs.ModNameHeaders).Logger(),
 		Compliance: complianceLog,
 	}
 }