+++ zetaclient/zetacore/client_vote.go
@@ -13,7 +13,6 @@ import (
 	observerclient "github.com/zeta-chain/node/x/observer/client/cli"
 	observertypes "github.com/zeta-chain/node/x/observer/types"
 	zctx "github.com/zeta-chain/node/zetaclient/context"
-	"github.com/zeta-chain/node/zetaclient/logs"
 )
 
 // PostVoteGasPrice posts a gas price vote. Returns txHash and error.
@@ -135,20 +134,15 @@ func (c *Client) PostVoteOutbound(
 	zetaTxHash, err := retry.DoTypedWithRetry(func() (string, error) {
 		return c.Broadcast(ctx, gasLimit, authzMsg, authzSigner)
 	})
-
 	if err != nil {
 		return "", ballotIndex, errors.Wrap(err, "unable to broadcast vote outbound")
 	}
 
 	go func() {
 		ctxForWorker := zctx.Copy(ctx, context.Background())
-
-		errMonitor := c.MonitorVoteOutboundResult(ctxForWorker, zetaTxHash, retryGasLimit, msg)
-		if errMonitor != nil {
-			c.logger.Error().
-				Err(err).
-				Str(logs.FieldMethod, "PostVoteOutbound").
-				Msg("failed to monitor vote outbound result")
+		err := c.MonitorVoteOutboundResult(ctxForWorker, zetaTxHash, retryGasLimit, msg)
+		if err != nil {
+			c.logger.Error().Err(err).Msg("failed to monitor vote outbound result")
 		}
 	}()
 
@@ -195,10 +189,7 @@ func (c *Client) PostVoteInbound(
 		ctxForWorker := zctx.Copy(ctx, context.Background())
 		errMonitor := c.MonitorVoteInboundResult(ctxForWorker, zetaTxHash, retryGasLimit, msg, monitorErrCh)
 		if errMonitor != nil {
-			c.logger.Error().
-				Err(errMonitor).
-				Str(logs.FieldMethod, "PostVoteInbound").
-				Msg("failed to monitor vote inbound result")
+			c.logger.Error().Err(errMonitor).Msg("failed to monitor vote inbound result")
 
 			if monitorErrCh != nil {
 				select {